producer/requirements.txt
requests==2.31.0
kafka-python==2.0.2
bs4==0.0.1
pandas==2.0.3